Move AdminSettings.vue to be closer to components which use it and correct paths accordingly
